Add text binding and fix assert function.
Fix the assert function to throw the error using the received error or an AssertError.
Parse enything in between '{' and '}' preventing binding more than once.
When a binding start tells the parser that it started and where and start an empty binding lexeme.
When a binding ends it validates the expression and finish the lexeme and the binding itself.